TF-571 Update token for interceptors dio
(cherry picked from commit ea598c20b8e0bb18db0c421e474ba22baea5585b)
